+++ tests/test_mapper.py
@@ -6,7 +6,7 @@ import pandas as pd
 
 from pyobo import Canonicalizer
 from pyobo.constants import XREF_COLUMNS
-from pyobo.xrefdb.xrefs_pipeline import get_graph_from_xref_df
+from pyobo.xrefdb.canonicalizer import get_graph_from_xref_df
 
 
 class TestCanonicalizer(unittest.TestCase):
